Move track page lineup actions to common
