+++ src/pages/track-page/TrackPageProvider.tsx
@@ -71,11 +71,12 @@ import {
 import { parseTrackRoute, TrackRouteParams } from 'utils/route/trackRouteParser'
 import { getTrackPageTitle, getTrackPageDescription } from 'utils/seo'
 
+import { tracksActions } from '../../common/store/pages/track/lineup/actions'
+
 import StemsSEOHint from './components/StemsSEOHint'
 import { OwnProps as DesktopTrackPageProps } from './components/desktop/TrackPage'
 import { OwnProps as MobileTrackPageProps } from './components/mobile/TrackPage'
 import * as trackPageActions from './store/actions'
-import { tracksActions } from './store/lineups/tracks/actions'
 import { TRENDING_BADGE_LIMIT } from './store/sagas'
 
 const getRemixParentTrackId = (track: Track | null) =>
