+++ src/pages/track-page/components/desktop/TrackPage.tsx
@@ -6,6 +6,7 @@ import { ID, CID } from 'common/models/Identifiers'
 import { LineupState } from 'common/models/Lineup'
 import { Track } from 'common/models/Track'
 import { User } from 'common/models/User'
+import { tracksActions } from 'common/store/pages/track/lineup/actions'
 import CoverPhoto from 'components/cover-photo/CoverPhoto'
 import Lineup from 'components/lineup/Lineup'
 import { LineupVariant } from 'components/lineup/types'
@@ -15,7 +16,6 @@ import SectionButton from 'components/section-button/SectionButton'
 import StatBanner from 'components/stat-banner/StatBanner'
 import GiantTrackTile from 'components/track/GiantTrackTile'
 import { TrackTileSize } from 'components/track/types'
-import { tracksActions } from 'pages/track-page/store/lineups/tracks/actions'
 import { getTrackDefaults, emptyStringGuard } from 'pages/track-page/utils'
 import { QueueItem } from 'store/queue/types'
