+++ src/pages/track-page/components/mobile/TrackPage.tsx
@@ -4,6 +4,7 @@ import { CID, ID } from 'common/models/Identifiers'
 import { LineupState } from 'common/models/Lineup'
 import { Track } from 'common/models/Track'
 import { User } from 'common/models/User'
+import { tracksActions } from 'common/store/pages/track/lineup/actions'
 import { OverflowAction } from 'common/store/ui/mobile-overflow-menu/types'
 import { HeaderContext } from 'components/header/mobile/HeaderContextProvider'
 import Lineup from 'components/lineup/Lineup'
@@ -16,7 +17,6 @@ import NavContext, {
 } from 'components/nav/store/context'
 import NetworkConnectivityMonitor from 'components/network-connectivity/NetworkConnectivityMonitor'
 import SectionButton from 'components/section-button/SectionButton'
-import { tracksActions } from 'pages/track-page/store/lineups/tracks/actions'
 import { getTrackDefaults, emptyStringGuard } from 'pages/track-page/utils'
 import { QueueItem } from 'store/queue/types'
