+++ src/pages/track-page/store/reducer.js
@@ -1,6 +1,8 @@
 import { asLineup } from 'common/store/lineup/reducer'
 import tracksReducer from 'pages/track-page/store/lineups/tracks/reducer'
 
+import { PREFIX as tracksPrefix } from '../../../common/store/pages/track/lineup/actions'
+
 import {
   SET_TRACK_ID,
   SET_TRACK_PERMALINK,
@@ -8,7 +10,6 @@ import {
   SET_TRACK_RANK,
   SET_TRACK_TRENDING_RANKS
 } from './actions'
-import { PREFIX as tracksPrefix } from './lineups/tracks/actions'
 
 const initialState = {
   trackId: null,
